Keep CPU busy when not paused
To get a decent time resolution on our measurements we need the CPU to
be clocked fast. Using an ondemand governor, this means that we need to
load the CPU before trying to do any measurements. So while unpaused, we
should do busy waiting whenever possible and when triggering, we should
start loading the CPU _before_ doing the actual trigger.