Terminate if access to message queue is lost
When the message queue is destroyed, libgpiod_pulsein will loop on
msgrcv() and hog the CPU. There is no way to recover from this situation
so just terminate with an error message in this case.

This situation can occur when a background task (e.g., a systemd
service) runs under a non-system user and uses libgpiod_pulsein. When
that user logs out from all of its sessions, logind in its default
configutation will destroy all remaining message queues of that user.

Refs adafruit/Adafruit_CircuitPython_DHT#50